% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhsc.rap.workshop.excersise.5-package.R
\docType{package}
\name{dhsc.rap.workshop.excersise.5-package}
\alias{dhsc.rap.workshop.excersise.5}
\alias{dhsc.rap.workshop.excersise.5-package}
\title{dhsc.rap.workshop.excersise.5: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{email@dhsc.gov.uk}

}
\keyword{internal}
