% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_ons_data}
\alias{download_ons_data}
\title{Download ONS data - retrieves ONS data using the ID specified in the config file}
\usage{
download_ons_data(config)
}
\arguments{
\item{config}{\itemize{
\item contains I/O directory structure, ONS url and id for desired data
}}
}
\value{
ons_data as a dataframe
}
\description{
Download ONS data - retrieves ONS data using the ID specified in the config file
}
