% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plot_summarise.R
\name{filter_data}
\alias{filter_data}
\title{Filter ONS data - cleas and filters the ONS data}
\usage{
filter_data(df)
}
\arguments{
\item{df}{\itemize{
\item the data set to be filtered
}}
}
\value{
a cleaned version of the data
}
\description{
Filter ONS data - cleas and filters the ONS data
}
