% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plot_summarise.R
\name{plot_data}
\alias{plot_data}
\title{Plot ONS data - create a ggplot graph of the data and save it to output}
\usage{
plot_data(df)
}
\arguments{
\item{df}{\itemize{
\item the data set to be plotted
}}
}
\description{
Plot ONS data - create a ggplot graph of the data and save it to output
}
