% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_plot_summarise.R
\name{summary_stats}
\alias{summary_stats}
\title{Summarise ONS data - calulate summary stat of the ons data and print the results}
\usage{
summary_stats(df)
}
\arguments{
\item{df}{\itemize{
\item the data set to be summarised
}}
}
\description{
Summarise ONS data - calulate summary stat of the ons data and print the results
}
